Account for non-zero exit codes, this'll happen when hosts are down
Also don't bother to sort the output anymore (as I did in d94e403a
Print multiping responses in order). It's the config output that has
to be sorted, not the actual status output. So to do this I need to
persist the return values across iterations, and then sort them.

Or just alter the plugin definition to call things in the order that
are likely to be slow -> fast.